Modified minicomponent_registry to only remove javascript files on registry build task as it was removing temporary js files that are needed in admin.

+++ jomres/libraries/jomres/classes/minicomponent_registry.class.php
@@ -105,13 +105,17 @@ class minicomponent_registry
 		$c->eraseAll();
 		
 		//delete js files in /jomres/temp dir
-		$javascript_files_in_temp_dir = scandir_getfiles( JOMRESCONFIG_ABSOLUTE_PATH . JRDS . JOMRES_ROOT_DIRECTORY . JRDS . 'temp' . JRDS, $extension = "js" );
-		foreach ( $javascript_files_in_temp_dir as $file )
+		
+		if (isset($_REQUEST['task']) && $_REQUEST['task'] == "rebuildregistry")
 			{
-			unlink( JOMRESCONFIG_ABSOLUTE_PATH . JRDS . JOMRES_ROOT_DIRECTORY . JRDS . 'temp' . JRDS . $file );
+			$javascript_files_in_temp_dir = scandir_getfiles( JOMRESCONFIG_ABSOLUTE_PATH . JRDS . JOMRES_ROOT_DIRECTORY . JRDS . 'temp' . JRDS, $extension = "js" );
+			foreach ( $javascript_files_in_temp_dir as $file )
+				{
+				unlink( JOMRESCONFIG_ABSOLUTE_PATH . JRDS . JOMRES_ROOT_DIRECTORY . JRDS . 'temp' . JRDS . $file );
+				}
 			}
 		}
-
+		
 	function save_registry_file()
 		{
 		$existed = false;
